// src/typings.d.ts
export type dataObj = {
    productCategory: string,
    prjCategory: string,
    prjName: string,
    prjManager: string,
    prjStartDate: string,
    prjEndDate: string,
    guid: string,
    projectStatus: string,
    department: string,
    logo: string,
    taskCount: number,
    taskDoneCount: number,
    taskDoingCount: number
}